// src/app/generated_services/model/showBeltDTO.ts
/**
 * Carlson Gracie Backend API
 * API for managing students, belts, graduation requirements, notices, frequencies and user authentication with Keycloak
 *
 * Contact: support@carlsongracie.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface ShowBeltDTO { 
    id?: string;
    color: string;
    orderIndex?: number;
    isForKids?: boolean;
}